less_noise: attempt to make feedback a little cleaner

+++ Tools/Apply_Link_Check.sh
@@ -3,7 +3,9 @@
 # Script in charge of auditing the released MD files in order to detect dead links
 
 cd ../Document
-rm ../link-check-result.out
+if test -f "../link-check-result.out"; then
+        rm ../link-check-result.out
+fi
 find . -name \*.md -exec markdown-link-check -q -c ../.markdownlinkcheck.json {} \; 1>../link-check-result.out 2>&1
 errors=`grep -c "ERROR:" ../link-check-result.out`
 content=`cat ../link-check-result.out`
@@ -13,4 +15,5 @@ then
     exit $errors
 else
     echo "[+] No error found by the Links validator."
+    rm ../link-check-result.out
 fi
