+++ Tools/Apply_Linter_Check.sh
@@ -3,7 +3,9 @@
 # Script in charge of auditing the released MD files with the linter policy defined at project level
 
 cd ..
-rm linter-result.out
+if test -f "linter-result.out"; then
+        rm linter-result.out
+fi
 markdownlint -c .markdownlint.json -o linter-result.out Document
 errors=`wc -m linter-result.out | cut -d' ' -f1`
 content=`cat linter-result.out`
@@ -14,4 +16,5 @@ then
     #exit $errors
 else
     echo "[+] No error found by the Linter."
+    rm linter-result.out
 fi
